Rename RenderLoop -> DefaultRenderLoop.
Make it overridable and register it in the windowing subsystem, so that
windowing subsystems can provide custom implementations.
